extract duplicated code to a new util
