+++ server/src/algorithms/tesserOCR.py
@@ -1,8 +1,8 @@
-import re
 from tesserocr import PyTessBaseAPI, PSM
-from lxml import etree
-from lxml import html
 from PIL import Image, ImageEnhance, ImageFilter
+
+from src.utils.parse_hocr import parse_hocr
+
 
 def preprocess_image(image):
     # Downscale to 200 DPI (assuming 300 DPI input)
@@ -14,6 +14,7 @@ def preprocess_image(image):
     image = image.filter(ImageFilter.MedianFilter())  # Noise removal
     image = image.point(lambda p: 255 if p > 128 else 0)  # Binarization
     return image
+
 
 def get_segment_hocr(page, config, segment_coordinates):
     """
@@ -33,6 +34,7 @@ def get_segment_hocr(page, config, segment_coordinates):
         hocr = api.GetHOCRText(0)
     return hocr
 
+
 def get_hocr(page, config):
     """
     Get the HOCR of an entire page.
@@ -49,27 +51,6 @@ def get_hocr(page, config):
         hocr = api.GetHOCRText(0)
     return hocr
 
-def is_left(point_a, point_b, point_c):
-    dx = 5
-    return ((point_b[0] + dx) - (point_a[0] + dx)) * (point_c[1] - point_a[1]) - (point_b[1] - point_a[1]) * (point_c[0] - (point_a[0] + dx)) >= 0
-
-def remove_extra_paragraphs(lines):
-    new_lines = [lines[0]]
-    start_coords = [(line[0]["box"][0], line[0]["box"][1]) for line in lines]
-
-    if len(set(x[0] for x in start_coords)) == 1:
-        # All vertical join all lines
-        for i in range(1, len(lines)):
-            line = lines[i]
-            new_lines[0].extend(line)
-    else:
-        for i in range(1, len(lines)):
-            if is_left(start_coords[0], start_coords[-1], start_coords[i]):
-                new_lines[-1].extend(lines[i])
-            else:
-                new_lines.append(lines[i])
-
-    return new_lines
 
 def get_structure(page, config, segment_box=None):
     """
@@ -86,7 +67,7 @@ def get_structure(page, config, segment_box=None):
     # Ensure config is a dict, use defaults if not
     if not isinstance(config, dict):
         config = {"psm": PSM.SINGLE_BLOCK, "lang": "por"}
-    
+
     with PyTessBaseAPI(psm=config.get("psm", PSM.SINGLE_BLOCK), lang=config.get("lang", "por"), oem=1) as api:
         api.SetImage(page)
         if segment_box:
@@ -103,67 +84,3 @@ def get_structure(page, config, segment_box=None):
         else:
             hocr = api.GetHOCRText(0)
         return parse_hocr(hocr, segment_box)
-
-def parse_hocr(hocr, segment_box):
-    """
-    Parse HOCR string into structured text data.
-
-    :param hocr: HOCR string from Tesseract.
-    :param segment_box: Bounding box of the segment, if applicable.
-    :return: List of lines, each containing words with text and coordinates.
-    """
-    p1 = re.compile(r"bbox((\s+\d+){4})")
-    p2 = re.compile(r"baseline((\s+[\d\.\-]+){2})")
-
-    hocr = etree.fromstring(hocr, html.XHTMLParser())
-
-    lines = []
-
-    for line in hocr.xpath('//*[@class="ocr_line"]'):
-        linebox = p1.search(line.attrib["title"]).group(1).split()
-        try:
-            baseline = p2.search(line.attrib["title"]).group(1).split()
-        except AttributeError:
-            baseline = [0, 0]
-        linebox = [float(i) for i in linebox]
-        baseline = [float(i) for i in baseline]
-
-        words = []
-
-        xpath_elements = './/*[@class="ocrx_word"]'
-        if not (line.xpath("boolean(" + xpath_elements + ")")):
-            # If there are no word elements present, switch to lines as elements
-            xpath_elements = "."
-
-        for word in line.xpath(xpath_elements):
-            rawtext = word.text_content().strip()
-            if rawtext == "":
-                continue
-
-            box = p1.search(word.attrib["title"]).group(1).split()
-
-            if segment_box:
-                box = [float(i) + segment_box[id % 2] for id, i in enumerate(box)]
-            else:
-                box = [float(i) for i in box]
-            b = polyval(baseline, (box[0] + box[2]) / 2 - linebox[0]) + linebox[3]
-
-            words.append({"text": rawtext, "box": box, "b": b})
-
-        if words:
-            lines.append(words)
-
-    if segment_box and lines:
-        lines = remove_extra_paragraphs(lines)
-
-    return lines
-
-def polyval(poly, x):
-    """
-    Evaluate a polynomial at a given x value.
-
-    :param poly: Polynomial coefficients [a, b].
-    :param x: x-value to evaluate.
-    :return: Evaluated value.
-    """
-    return x * poly[0] + poly[1]
